requirements.txt
requests==2.21.0
python-dotenv==0.10.2
instabot==0.44.0
Pillow==6.0.0